chart configurator works again
