+++ client/printers/util.js
@@ -1,0 +1,6 @@
+export function getDefaultMeta () {
+	return {
+		label: undefined,
+		question: undefined
+	};
+}
