+++ views/overview.html
@@ -9,7 +9,7 @@
 	</div>
 	<div data-o-grid-colspan="10">
 		<div class="chart" data-keen-alias="members/daily">
-			{{>chart__configurator}}
+			{{>chart-configurator}}
 			<div class="chart__printer chart-loading core-chart"></div>
 		</div>
 	</div>
